requirements.txt
streamlit>=1.20.0
pillow>=9.0.0
pytesseract>=0.3.10
streamlit-image-coordinates>=0.1.6
streamlit-tesseract-scanner>=0.1.5
